requirements.txt
Flask==1.0.2
gunicorn==19.9.0
Pillow=7.0.0
